+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning][].
 
 * Displaying the firmware version on the Device view
 * Dynamic keyboard preview. Halo75 and Halo65 support with correct layout.
+* Help message for Sidelight. It may seem like the settings don't apply, but it's more likely that your keyboard isn't charged to the max
 
 ## Fixed